+++ frontend/src/App_private.tsx
@@ -32,7 +32,7 @@ const App_private: React.FC = () => {
     <div className='App_private'>
       <h2>Private APP</h2>
 
-      <Link href='/home'>
+      <Link href='/'>
         <a>Home</a>
       </Link>
       <br />
@@ -60,8 +60,8 @@ const App_private: React.FC = () => {
       <br></br>
       <br></br>
       <Switch>
-        <Route path='/' component={Welcome} />
-        <Route path='/home' component={Home} />
+        <Route path='/welcome' component={Welcome} />
+        <Route path='/' component={Home} />
         <Route path='/chat' component={PrivateMessage} />
         <Route path='/testUser' component={FindUser} />
         <Route path='/game' component={Game} />
