+++ frontend/src/auth/Login/logUser.tsx
@@ -1,4 +1,9 @@
+import { useLocation } from 'wouter'
+import React, { useEffect } from 'react'
+
 export async function LogUser(mail: string, password: string) {
+  // const [, setLocation] = useLocation()
+  
   try {
     const response = await fetch('http://127.0.0.1:3000/api/auth/login', {
       method: 'POST',
@@ -6,7 +11,7 @@ export async function LogUser(mail: string, password: string) {
         'Content-Type': 'application/json'
       },
       credentials: 'include',
-
+      
       body: JSON.stringify({
         mail,
         password
@@ -14,9 +19,10 @@ export async function LogUser(mail: string, password: string) {
     })
     if (response.ok) {
       const userData = await response.json()
-      window.location.href = 'http://127.0.0.1:5173' // replace with setLocation('/', { replace: true }) ?
       return userData
     }
+    else 
+      return null
   } catch (error) {
     throw error
   }
